feat: move EntityHighlighter to folder, create colors and helpers
