+++ src/components/EntityHighlighter/helpers.js
@@ -1,0 +1,15 @@
+export function hashString(str) {
+    if (str.length === 0) {
+        return 0;
+    }
+
+    let hash = 0;
+    
+    for (let i = 0; i < str.length; i++) {
+        const char = str.charCodeAt(i);
+        hash = (hash << 5) - hash + char;
+        hash &= hash; // Convert to 32bit integer
+    }
+
+    return hash > 0 ? hash : -hash;
+}
