+++ src/components/EntityHighlighter/index.js
@@ -1,0 +1,3 @@
+import EntityHighlighter from "./EntityHighlighter";
+
+export default EntityHighlighter;
